requirements.txt
<<<<<<< HEAD
<<<<<<< HEAD
streamlit
google-cloud-bigquery
pandas
=======
streamlit
google-cloud-bigquery
pandas
>>>>>>> origin/main
=======
streamlit
google-cloud-bigquery
pandas
>>>>>>> origin/main
db-dtypes